requirements.txt
transformers==4.40.0
torch==2.2.0
accelerate
bitsandbytes==0.43.0
pdfplumber
keybert
spacy
jinja2==3.1.3
pyyaml==6.0.1
fastapi
uvicorn
azure-storage-blob
tiktoken
ipython